fix(spanner): widen a retry-count comparison in a scheduler test helper
Follow-up to widening ProcessTracker::retry_count to i64. `walk_pcr_ladder`
and the constant feeding it were still typed i32, so the call to
get_pcr_payments_retry_schedule_time no longer type-checked. `cargo build`
does not compile test targets, so this never reached the image build, but it
fails `cargo test`.

The sibling fix in api_key_expiry.rs, which did break the image build, is
1e77414f6c.

Verified with the image build's exact feature set, plus test targets:

    cargo check --workspace --all-targets --no-default-features \
        --features release --features v1,redis-rs --features spanner

Clean. That is the combination the earlier checks should have used: a
default-feature check leaves the `release`-only gates -- email,
dynamic_routing, deja, keymanager_mtls, vergen, gcp_kms -- uncompiled, which
is why api_key_expiry.rs was missed until Jenkins hit it.

+++ crates/scheduler/src/utils.rs
@@ -502,7 +502,7 @@ mod tests {
     /// slot is the delay before one retry, so `N` slots describe the charge plus `N` retries.
     fn walk_pcr_ladder(
         mapping: &process_data::RetryMapping,
-        attempts_already_made: i32,
+        attempts_already_made: i64,
     ) -> (usize, i32) {
         let mut retry_count = attempts_already_made;
         let mut elapsed = 0;
@@ -527,7 +527,7 @@ mod tests {
     #[test]
     fn test_pcr_retry_ladder_leaves_thirteen_retries_to_us_ending_on_day_twenty_eight() {
         const DAY: i32 = 24 * 60 * 60;
-        const BILLING_CONNECTOR_RETRY_THRESHOLD: i32 = 2;
+        const BILLING_CONNECTOR_RETRY_THRESHOLD: i64 = 2;
 
         let mapping = process_data::RetryMapping {
             start_after: DAY,
